refactor jwt generator
