+++ src/data/repository/auth/jwt/generator.rs
@@ -1,12 +1,11 @@
-use std::{collections::BTreeMap, env};
-
 use chrono::Duration;
 use hmac::{Hmac, Mac};
 use jwt::SignWithKey;
 use sha2::Sha256;
+use std::{collections::BTreeMap, env};
 
 use super::{
-    objects::{JwtObject, JwtResult},
+    objects::{JwtGeneratorError, JwtObject, JwtResult},
     JwtGenerator,
 };
 
@@ -90,40 +89,6 @@ impl JwtGenerator for JwtObject {
                 log::error!("Failed to sign with key");
                 Err(JwtGeneratorError::SignWithKey)
             }
-        }
-    }
-}
-
-#[derive(Debug)]
-pub enum JwtGeneratorError {
-    InvalidEnvSecret,
-    DurationOutOfBound,
-    TimeCreationError,
-    SignWithKey,
-    CreateKey,
-}
-
-impl std::fmt::Display for JwtGeneratorError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            JwtGeneratorError::InvalidEnvSecret => write!(f, "Invalid env secret"),
-            JwtGeneratorError::TimeCreationError => write!(f, "Time creation error"),
-            JwtGeneratorError::SignWithKey => write!(f, "Sign with key error"),
-            JwtGeneratorError::CreateKey => write!(f, "Create key error"),
-            JwtGeneratorError::DurationOutOfBound => write!(f, "Duration out of bound"),
-        }
-    }
-}
-
-impl PartialEq for JwtGeneratorError {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (JwtGeneratorError::InvalidEnvSecret, JwtGeneratorError::InvalidEnvSecret) => true,
-            (JwtGeneratorError::DurationOutOfBound, JwtGeneratorError::DurationOutOfBound) => true,
-            (JwtGeneratorError::TimeCreationError, JwtGeneratorError::TimeCreationError) => true,
-            (JwtGeneratorError::SignWithKey, JwtGeneratorError::SignWithKey) => true,
-            (JwtGeneratorError::CreateKey, JwtGeneratorError::CreateKey) => true,
-            _ => false,
         }
     }
 }
