+++ src/data/repository/auth/jwt/tests.rs
@@ -6,7 +6,9 @@ mod test_generator {
     use jwt::{Error, Header, Token, VerifyWithKey};
     use sha2::Sha256;
 
-    use crate::utils::jwt_util::{generator::JwtGeneratorError, objects::JwtObject, JwtGenerator};
+    use crate::data::repository::jwt::{
+        generator::JwtGeneratorError, objects::JwtObject, JwtGenerator,
+    };
 
     const EXPECTED_UUID: &str = "expected_uuid";
     const EXPECTED_USERNAME: &str = "expected_username";
