+++ src/utils/jwt_util/mod.rs
@@ -1,25 +1,8 @@
-use self::{
-    decoder::JwtDecoderError,
-    generator::JwtGeneratorError,
-    objects::{JwtDecoderResult, JwtResult},
-};
+use self::decoder::{JwtDecoderError, JwtDecoderResult};
 
 pub mod decoder;
-pub mod generator;
-pub mod objects;
-mod tests;
 
-#[async_trait]
-pub trait JwtGenerator {
-    async fn generate(&self) -> Result<JwtResult, JwtGeneratorError>;
-    async fn generate_access(&self) -> Result<String, JwtGeneratorError>;
-    async fn generate_refresh(&self) -> Result<String, JwtGeneratorError>;
-    async fn generate_token(
-        &self,
-        secret: &[u8],
-        exp_days: i64,
-    ) -> Result<String, JwtGeneratorError>;
-}
+mod tests;
 
 pub trait JwtDecoder {
     fn decode_refresh(&self) -> Result<JwtDecoderResult, JwtDecoderError>;
