+++ src/utils/jwt_util/tests/mod.rs
@@ -1,3 +1,2 @@
 mod mapper;
 mod test_decoder;
-mod test_generator;
